# [mypyc] Make multiple assignment faster(1st stage)