relates mypyc/mypyc#729

This PR makes simple cases like `y, x = x, y` faster.

I tend to do some future improvement based on this PR(for example, supporting `ListExpr` as @JukkaL suggested), nevertheless, this PR itself should cover the most frequent use case.