@TH3CHARLie 2020-10-10
Some imprecise benchmark results:

Before:
```
............
interpreted: 0.167262s (avg of 6 iterations; stdev 0.61%)
compiled:    0.163983s (avg of 6 iterations; stdev 0.48%)

compiled is 1.020x faster
```
After:
```
running multiple_assignment
............
interpreted: 0.164994s (avg of 6 iterations; stdev 0.52%)
compiled:    0.055589s (avg of 6 iterations; stdev 0.69%)

compiled is 2.968x faster
```